+++ UnitTests/SettingsTest.cs
@@ -34,14 +34,14 @@ EqualsInValue=We can have '=' in the value
 
             LinkedList<QuickFix.SettingsDictionary> foo = settings.Get("FOO");
             Assert.That(foo.Count, Is.EqualTo(1));
-            Assert.That(foo.First.Value.GetLong("bar"), Is.EqualTo(24));
+            Assert.That(foo.First!.Value.GetLong("bar"), Is.EqualTo(24));
             Assert.That(foo.First.Value.GetString("baz"), Is.EqualTo("moo"));
             Assert.That(foo.First.Value.GetString("baz"), Is.EqualTo("moo"));
             Assert.That(foo.First.Value.Count, Is.EqualTo(2));
 
             LinkedList<QuickFix.SettingsDictionary> oren = settings.Get("OREN");
             Assert.That(oren.Count, Is.EqualTo(2));            
-            Assert.That(oren.First.Value.Count, Is.EqualTo(1));
+            Assert.That(oren.First!.Value.Count, Is.EqualTo(1));
             Assert.That(oren.First.Value.GetString("Nero"), Is.EqualTo("TW"));
             oren.RemoveFirst();
             Assert.That(oren.First.Value.Count, Is.EqualTo(2));
@@ -50,7 +50,7 @@ EqualsInValue=We can have '=' in the value
 
             LinkedList<QuickFix.SettingsDictionary> nero = settings.Get("NERO");
             Assert.That(nero.Count, Is.EqualTo(1));
-            Assert.That(nero.First.Value.Count, Is.EqualTo(5));
+            Assert.That(nero.First!.Value.Count, Is.EqualTo(5));
             Assert.That(nero.First.Value.GetString("WINDIR"), Is.EqualTo("D:\\This Is\\A-Directory\\Connamara\\"));
             Assert.That(nero.First.Value.GetString("UNIXDIR"), Is.EqualTo("/home/mgatny/This Is/A Directory/ok/"));
             Assert.That(nero.First.Value.GetString("WINFILE"), Is.EqualTo("D:\\Program Files\\Tomcat 4.1\\webapps\\mgatny\\WEB-INF\\connamara.cfg"));
@@ -70,14 +70,14 @@ what=huh";
 
             LinkedList<QuickFix.SettingsDictionary> byLower = settings.Get("foo");
             Assert.AreEqual(1, byLower.Count);
-            Assert.AreEqual(2, byLower.First.Value.Count);
+            Assert.AreEqual(2, byLower.First!.Value.Count);
             Assert.AreEqual("uno", byLower.First.Value.GetString("one"));
             Assert.AreEqual("dos", byLower.First.Value.GetString("two"));
 
             // too lazy to write a QuickFix.Dictionary#Equals method (which would only be used by this test)
             LinkedList<QuickFix.SettingsDictionary> byUpper = settings.Get("FOO");
             Assert.AreEqual(byLower.Count, byUpper.Count);
-            Assert.AreEqual(byLower.First.Value.Count, byUpper.First.Value.Count);
+            Assert.AreEqual(byLower.First.Value.Count, byUpper.First!.Value.Count);
             Assert.AreEqual(byUpper.First.Value.GetString("one"), byUpper.First.Value.GetString("one"));
             Assert.AreEqual(byUpper.First.Value.GetString("two"), byUpper.First.Value.GetString("two"));
         }